refactor: :recycle: update Tabs component structure
