--- src/components/@base/Tabs/TabsButton.stories.tsx
@@ -1,33 +1,0 @@
-import DesignSystem from '@/utils/designSystem';
-import { StoryObjWithCSSProp } from '@/utils/types';
-import type { Meta } from '@storybook/react';
-import TabsButton from './TabsButton';
-
-const meta: Meta<Omit<typeof TabsButton, 'defaultProps'>> = {
-  component: TabsButton,
-};
-
-export default meta;
-
-type Story = StoryObjWithCSSProp<typeof TabsButton>;
-
-export const DefaultTabsButton: Story = {
-  args: {
-    bgColor: DesignSystem.Color.primary['yellow'],
-    children: 'mybutton',
-  },
-  render: (args) => <TabsButton {...args} />,
-};
-
-export const TabsButtonWithCssProps: Story = {
-  args: {
-    ...DefaultTabsButton.args,
-    css: {
-      borderRadius: 40,
-      padding: 10,
-      border: `1px solid ${DesignSystem.Color.text.gray}`,
-      color: DesignSystem.Color.text.gray,
-    },
-  },
-  render: (args) => <TabsButton {...args} />,
-};
