+++ src/components/@base/Tabs/components/Button.tsx
@@ -1,6 +1,6 @@
-import { DEFAULT_PROPS_PREFIX } from '@/utils/constants';
 import { CSSProperties, HTMLAttributes, useContext } from 'react';
-import { TabValueType, TabsContext } from './TabsMain';
+import { TabValueType, TabsContext } from '../Tabs';
+import { DEFAULT_PROPS_PREFIX } from '../utils';
 
 export interface TabsButtonProps extends HTMLAttributes<HTMLButtonElement> {
   value: TabValueType;
@@ -22,7 +22,7 @@ function TabsButton({
       css={{ backgroundColor, color: 'inherit' }}
       onClick={(e) => {
         onTabChange(value);
-        onClick && onClick(e);
+        onClick?.(e);
       }}
       {...props}
     >
