+++ src/components/@base/Tabs/index.ts
@@ -1,6 +1,6 @@
-import TabsBody from './TabsBody';
-import TabsButton from './TabsButton';
-import TabsMain from './TabsMain';
+import TabsBody from './components/Body';
+import TabsButton from './components/Button';
+import TabsMain from './Tabs';
 
 const Tabs = Object.assign(TabsMain, {
   Body: TabsBody,
