+++ src/utils/constants.ts
@@ -1,7 +1,5 @@
 import DesignSystem from './designSystem';
 
-export const DEFAULT_PROPS_PREFIX = '_type';
-
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 function getObjectKeys(obj: any, parentKey = ''): string[] {
   return Object.keys(obj).flatMap((key) => {
